requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
recipe-scrapers==14.51.0
python-dotenv==1.0.0